requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
djangorestframework==3.15.1
gunicorn==21.2.0
idna==3.6
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
requests==2.31.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
